genericize TestSpecOutput
